refactor: change according to pull request comment
1. add mol schema files of omni_lock and xudt_rce_mol into src/types/schemas;
2. unreachable!() -> todo!()
3. remove to_address_payload() and type_hash from OmniLockConfig
4. move OmniLockScriptSigner to signer.rs, move OmniLockUnlocker to unlocker.rs
5. define enum for omnilock id flags
