+++ src/types/schemas/omni_lock.mol
@@ -1,0 +1,16 @@
+import xudt_rce;
+
+array Auth[byte; 21];
+
+table Identity {
+    identity: Auth,
+    proofs: SmtProofEntryVec,
+}
+option IdentityOpt (Identity);
+
+// the data structure used in lock field of witness
+table OmniLockWitnessLock {
+    signature: BytesOpt,
+    omni_identity: IdentityOpt,
+    preimage: BytesOpt,
+}
