+++ src/unlock/mod.rs
@@ -4,11 +4,12 @@ mod unlocker;
 
 pub use signer::{
     generate_message, AcpScriptSigner, ChequeAction, ChequeScriptSigner, MultisigConfig,
-    ScriptSignError, ScriptSigner, SecpMultisigScriptSigner, SecpSighashScriptSigner,
+    OmniLockScriptSigner, ScriptSignError, ScriptSigner, SecpMultisigScriptSigner,
+    SecpSighashScriptSigner,
 };
 pub use unlocker::{
-    fill_witness_lock, reset_witness_lock, AcpUnlocker, ChequeUnlocker, ScriptUnlocker,
-    SecpMultisigUnlocker, SecpSighashUnlocker, UnlockError,
+    fill_witness_lock, reset_witness_lock, AcpUnlocker, ChequeUnlocker, OmniLockUnlocker,
+    ScriptUnlocker, SecpMultisigUnlocker, SecpSighashUnlocker, UnlockError,
 };
 
-pub use omni_lock::{OmniLockConfig, OmniLockScriptSigner, OmniLockUnlocker};
+pub use omni_lock::{IdentityFlags, OmniLockConfig};
